Move variable __gmt_version__ to pygmt.clib to avoid cyclic-import errors
