+++ pygmt/clib/__init__.py
@@ -6,3 +6,6 @@ Pythonic interface. Access to the C library is done through ctypes.
 """
 
 from pygmt.clib.session import Session
+
+with Session() as lib:
+    __gmt_version__ = lib.info["version"]
